requirements.txt
pandas
requests
boto3
certifi
psycopg2-binary
dbt-core
dbt-postgres
